requirements.txt
amqp==2.5.2
appdirs==1.4.3
attrs==19.3.0
billiard==3.5.0.5
black==19.10b0
celery==4.2.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.5
django-appconf==1.0.3
django-compressor==2.3
django-countries==5.3.1
django-hamlpy==1.2
django-qurl-templatetag==0.0.13
django-redis==4.11.0
django-sass-processor==0.7.3
django-taggit==1.1.0
django-timezone-field==4.0
django-widget-tweaks==1.4.5
djangorestframework==3.8.2
drf-yasg==1.17.0
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
iso8601==0.1.12
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
libsass==0.19.4
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.3
pathspec==0.8.0
phonenumbers==8.12.0
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.1
pytz==2019.3
rapidpro-dash==1.4.5
rapidpro-python==2.6.1
raven==6.10.0
rcssmin==1.0.6
redis==3.4.1
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sentry-sdk==0.14.3
six==1.14.0
smartmin==2.2.0
-e git+https://github.com/jazzband/sorl-thumbnail.git@b6358d234d7de3927a2666a2a5ab3d7870c0e1d3#egg=sorl_thumbnail
sqlparse==0.3.1
-e git+https://github.com/Alir3z4/python-stop-words.git@7b30f5b953ef02b62dfb4b8a158ea5f1218d11e7#egg=stop_words
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
